- [config] Fix formatting. - [config] Enable slidingExpiration to the default <authentication> config. - [net] Timeout setting for clients is an integer. - [net] There is no need of a real Timer to timeout read operations so now use the lightweight version. Now fewer system resources should be needed. - [GlobalUI] Removed some unused code. - [GlobalUI] Redirect to default.aspx when closing the session, the Forms Authenticator should do the rest of the work. - [GlobalUI] If we are authenticated, but session data is not consistent, then close session.

+++ trunk/SharpWebMail/controls/src/CTNSimplePOP3Client.cs
@@ -49,7 +49,7 @@ namespace anmar.SharpWebMail
 		/// <param name="user"></param>
 		/// <param name="pass"></param>
 		/// <param name="timeout"></param>
-		public CTNSimplePOP3Client( System.String host, System.Int32 port, System.String user, System.String pass, System.Double timeout ) : base(host, port, user, pass, timeout) {
+		public CTNSimplePOP3Client( System.String host, System.Int32 port, System.String user, System.String pass, long timeout ) : base(host, port, user, pass, timeout) {
 			this.commandEnd = "\r\n";
 			this.responseEnd = "\r\n.\r\n";
 			this.responseEndSL = "\r\n";
